Sync App List (Launcher) folders. This CL also: * Makes AppListSyncableService observer AppListModel instead of its   item list (and adds relevant observer methods to the model). * Protects Add/Remove/Delete methods in AppListItemList, making   AppListModel solely responsible for adding and removing items   from the primary list or any folders. * Adds folder support to AppListModel. * Fixes some edge cases in the app list sync code exposed in testing.
Patchset 5+ depends on https://codereview.chromium.org/148403007/

BUG=335761
For sync integration test changes:
TBR=zea@chromium.org

Review URL: https://codereview.chromium.org/147743002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@248079 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 050240d2ba58a73b2e03c073e03bb33e831d60d4

+++ ui/app_list/app_list_model.cc
@@ -43,12 +43,42 @@ void AppListModel::SetStatus(Status status) {
 }
 
 AppListItem* AppListModel::FindItem(const std::string& id) {
-  return item_list_->FindItem(id);
+  AppListItem* item = item_list_->FindItem(id);
+  if (item)
+    return item;
+  for (size_t i = 0; i < item_list_->item_count(); ++i) {
+    AppListItem* child_item = item_list_->item_at(i)->FindChildItem(id);
+    if (child_item)
+      return child_item;
+  }
+  return NULL;
+}
+
+AppListFolderItem* AppListModel::FindFolderItem(const std::string& id) {
+  AppListItem* item = item_list_->FindItem(id);
+  if (item && item->GetItemType() == AppListFolderItem::kItemType)
+    return static_cast<AppListFolderItem*>(item);
+  DCHECK(!item);
+  return NULL;
 }
 
 void AppListModel::AddItem(AppListItem* item) {
+  DCHECK(!item->IsInFolder());
   DCHECK(!item_list()->FindItem(item->id()));
   AddItemToItemList(item);
+}
+
+void AppListModel::AddItemToFolder(AppListItem* item,
+                                   const std::string& folder_id) {
+  if (folder_id.empty()) {
+    AddItem(item);
+    return;
+  }
+  DCHECK(!item->IsInFolder() || item->folder_id() == folder_id);
+  DCHECK(item->GetItemType() != AppListFolderItem::kItemType);
+  AppListFolderItem* dest_folder = FindOrCreateFolderItem(folder_id);
+  DCHECK(!dest_folder->item_list()->FindItem(item->id()));
+  AddItemToFolderItem(dest_folder, item);
 }
 
 const std::string& AppListModel::MergeItems(const std::string& target_item_id,
@@ -65,7 +95,7 @@ const std::string& AppListModel::MergeItems(const std::string& target_item_id,
   if (target_item->GetItemType() == AppListFolderItem::kItemType) {
     AppListFolderItem* target_folder =
         static_cast<AppListFolderItem*>(target_item);
-    AddItemToFolder(target_folder, source_item_ptr.release());
+    AddItemToFolderItem(target_folder, source_item_ptr.release());
     return target_folder->id();
   }
 
@@ -81,28 +111,81 @@ const std::string& AppListModel::MergeItems(const std::string& target_item_id,
   AddItemToItemList(new_folder);
 
   // Add the items to the new folder.
-  AddItemToFolder(new_folder, target_item_ptr.release());
-  AddItemToFolder(new_folder, source_item_ptr.release());
+  AddItemToFolderItem(new_folder, target_item_ptr.release());
+  AddItemToFolderItem(new_folder, source_item_ptr.release());
 
   return new_folder->id();
 }
 
+void AppListModel::MoveItemToFolder(AppListItem* item,
+                                    const std::string& folder_id) {
+  if (item->folder_id() == folder_id)
+    return;
+
+  // First, get or create the destination folder.
+  AppListFolderItem* dest_folder = FindOrCreateFolderItem(folder_id);
+
+  // Second, remove the item from its current location.
+  scoped_ptr<AppListItem> item_ptr;
+  if (item->IsInFolder()) {
+    AppListFolderItem* cur_folder = FindFolderItem(item->folder_id());
+    DCHECK(cur_folder);
+    item_ptr = RemoveItemFromFolder(cur_folder, item);
+  } else {
+    item_ptr = item_list_->RemoveItem(item->id());
+  }
+  DCHECK(item_ptr);
+  DCHECK(!item_ptr->IsInFolder());
+
+  // Finally, add the item to the new location and trigger observers.
+  if (dest_folder) {
+    AddItemToFolderItem(dest_folder, item_ptr.release());
+  } else {
+    item_list_->AddItem(item_ptr.release());
+    FOR_EACH_OBSERVER(AppListModelObserver,
+                      observers_,
+                      OnAppListItemUpdated(item));
+  }
+}
+
 void AppListModel::SetItemPosition(AppListItem* item,
                                    const syncer::StringOrdinal& new_position) {
-  item_list_->SetItemPosition(item, new_position);
-  // Note: this will trigger OnListItemMoved which will signal observers.
-  // (This is done this way because some View code still moves items within
-  // the item list directly).
+  if (!item->IsInFolder()) {
+    item_list_->SetItemPosition(item, new_position);
+    // Note: this will trigger OnListItemMoved which will signal observers.
+    // (This is done this way because some View code still moves items within
+    // the item list directly).
+    return;
+  }
+  AppListFolderItem* folder = FindFolderItem(item->folder_id());
+  DCHECK(folder);
+  folder->item_list()->SetItemPosition(item, new_position);
+  FOR_EACH_OBSERVER(AppListModelObserver,
+                    observers_,
+                    OnAppListItemUpdated(item));
 }
 
 void AppListModel::DeleteItem(const std::string& id) {
   AppListItem* item = FindItem(id);
   if (!item)
     return;
+  if (!item->IsInFolder()) {
+    DCHECK_EQ(0u, item->ChildItemCount())
+        << "Invalid call to DeleteItem for item with children: " << id;
+    FOR_EACH_OBSERVER(AppListModelObserver,
+                      observers_,
+                      OnAppListItemWillBeDeleted(item));
+    item_list_->DeleteItem(id);
+    return;
+  }
+  AppListFolderItem* folder = FindFolderItem(item->folder_id());
+  DCHECK(folder) << "Folder not found for item: " << item->ToDebugString();
+  scoped_ptr<AppListItem> child_item = RemoveItemFromFolder(folder, item);
+  DCHECK_EQ(item, child_item.get());
   FOR_EACH_OBSERVER(AppListModelObserver,
                     observers_,
                     OnAppListItemWillBeDeleted(item));
-  item_list_->DeleteItem(id);
+  child_item.reset();  // Deletes item.
 }
 
 // Private methods
@@ -122,16 +205,46 @@ syncer::StringOrdinal AppListModel::GetLastItemPosition() {
   return item_list_->item_at(count - 1)->position();
 }
 
+AppListFolderItem* AppListModel::FindOrCreateFolderItem(
+    const std::string& folder_id) {
+  if (folder_id.empty())
+    return NULL;
+
+  AppListFolderItem* dest_folder = FindFolderItem(folder_id);
+  if (!dest_folder) {
+    dest_folder = new AppListFolderItem(folder_id);
+    dest_folder->set_position(GetLastItemPosition().CreateAfter());
+    AddItemToItemList(dest_folder);
+  }
+  return dest_folder;
+}
+
 void AppListModel::AddItemToItemList(AppListItem* item) {
+  DCHECK(!item->IsInFolder());
   item_list_->AddItem(item);
   FOR_EACH_OBSERVER(AppListModelObserver,
                     observers_,
                     OnAppListItemAdded(item));
 }
 
-void AppListModel::AddItemToFolder(AppListFolderItem* folder,
-                                   AppListItem* item) {
+void AppListModel::AddItemToFolderItem(AppListFolderItem* folder,
+                                       AppListItem* item) {
   folder->item_list()->AddItem(item);
+  item->set_folder_id(folder->id());
+  FOR_EACH_OBSERVER(AppListModelObserver,
+                    observers_,
+                    OnAppListItemUpdated(item));
+}
+
+scoped_ptr<AppListItem> AppListModel::RemoveItemFromFolder(
+    AppListFolderItem* folder,
+    AppListItem* item) {
+  DCHECK_EQ(item->folder_id(), folder->id());
+  scoped_ptr<AppListItem> result = folder->item_list()->RemoveItem(item->id());
+  if (folder->item_list()->item_count() == 0)
+    DeleteItem(folder->id());
+  item->set_folder_id("");
+  return result.Pass();
 }
 
 }  // namespace app_list
